Allow slides to receive individual themes
Files data_dir/<theme_name>.slide.{start,end}.tex form the slide's
theme. They surround the current slide's line. So, supposing we have the
following lines in a slide,

This is the first line of a slide
This is the second line of a slide
This is the third line of a slide
And music continues like this

Thus, the method texify() from this slide will generate:

\begin{frame}
% <theme_name>.slide.start.tex are put here
This is the first line of a slide\\
This is the second line of a slide\\
This is the third line of a slide\\
And music continues like this\\
% <theme_name>.slide.end.tex are put here
\end{frame}