Add really smart wrap line method
Old method will just wrap lines letting what is exceeding the max col
to the next line. However, in a slide, if a line does not fit, it's
better to split it in two lines of nearly the same size.

method slide._smart_wrap() will try do split line in n lines, where n is
the number of lines necessary when calculating without considering that
a word should not be split. Then it will search forward and backward for
spaces to decide where to really split the string. If the biggest string
is still shorter than the maximum length allowed, then this is
preferred. Otherwise, the shortest one is returned. Next iteration will
divide the string in n-1 pieces and so on.