Allow lines of a verses to be re-organized
When a slide is finished (meaning the input text was already parsed and
all the lines that should be in a Verse were already added), then allow
the lines to be reorganized among Slides.

Basically what is allowed is to remove lines from the end of a slide and
prepend them on the subsequent slide. This is needed in order to allow
breaking the slide in other positions, so we don't have orphan lines,
which is done in a coming commit.