Add grammar support to block
A block in a slide is a piece of the lyric that should always be
kept together. It will not be split neither by smart or dumb
splitters (a warning will be generated). Also, the subsequent line is
a string that will be left in front of this block (centered vertically).
This is useful for telling the number of times a block should be
repeated. For example:

	This is a title

	This is the first line of the first verse
	though it could be split. But, never mind
	Let's start another verse

	This is another verse. Next line we'll repeat
	\
	|Repeat this line
	|And this, twice (see above)
	2x

The second verse contains a line, a blank line and a block that will
have a marker (2x) in front of it.

As could be seen, the syntax for repetition block is:

 * Mark lines of a block starting with a "|"
 * After a block is started, first line without "|" is considered to be
   the marker, and block is closed.

This implement only the grammar part. Nothing is done except from adding
the line, yet.